add core-primitive crate
