+++ Cargo.toml
@@ -12,6 +12,8 @@ members = [
 	"pallets/*",
 	"precompiles/*",
 
+	"core-primitives",
+
 	"chain-extensions/dapps-staking",
 	"chain-extensions/pallet-assets",
 	"chain-extensions/xvm",
@@ -273,6 +275,8 @@ pallet-xvm = { path = "./pallets/pallet-xvm", default-features = false }
 pallet-xcm = { path = "./pallets/pallet-xcm", default-features = false }
 
 xcm-primitives = { path = "./primitives/xcm", default-features = false }
+
+astar-core-primitives = { path = "./core-primitives", default-features = false }
 
 pallet-evm-precompile-assets-erc20 = { path = "./precompiles/assets-erc20", default-features = false }
 pallet-evm-precompile-sr25519 = { path = "./precompiles/sr25519", default-features = false }
